refactor (tests): split PasteServiceExecutionTests and extract mocks
- Created PasteServiceExecutionMocks.swift to house mock classes and helper types.
- Renamed NoopFallbackExecutor to PasteServiceNoopFallbackExecutor to avoid naming collisions with smoke tests.
- Removed approximately 200 lines of mock declarations from PasteServiceExecutionTests.swift to improve readability.
